キー入力問題解決。Releaseもビルドしたら治った。リザルト仮完成。タイトルも仮完成。

+++ tetris/DirectX.cpp
@@ -123,7 +123,7 @@ void DirectX::UpdateKeyState() {
 	}
 }
 
-bool DirectX::GetKeyState(BYTE KeyNumber) {
+bool DirectX::GetKeyState(int KeyNumber) {
 	if (KeyState[KeyNumber] & 0x80) {
 		return true;
 	}
